# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CofradeNet API
  description: |
    API REST de la plataforma CofradeNet. Gestiona hermandades, bandas, procesiones,
    usuarios cofrades y el panel de administración de la plataforma.

    ## Autenticación
    La mayoría de los endpoints de escritura requieren un token JWT Bearer.
    Obtén el token con `POST /auth/login` e inclúyelo en la cabecera:
    `Authorization: Bearer <token>`

    ## Roles disponibles
    | Rol | Descripción |
    |-----|-------------|
    | `cofrade` | Usuario registrado. Puede seguir entidades, crear publicaciones y gestionar su perfil. |
    | `hermandad` | Administra el perfil de su hermandad y sus procesiones. |
    | `banda` | Administra el perfil de su banda y su agenda de actuaciones. |
    | `admin` | Acceso completo a la plataforma. |
  version: 1.0.0
  contact:
    name: Equipo CofradeNet
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /usuarios/favoritos/banda/{id}:
    post:
      tags: [usuarios]
      summary: Añadir / quitar banda de favoritos (toggle)
      operationId: toggleFavoritoBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Estado de favorito actualizado.

  # ── HERMANDADES ───────────────────────────
...
  /bandas:
    get:
      tags: [bandas]
      summary: Listar todas las bandas
      operationId: listarBandas
      security: []
      responses:
        '200':
          description: Lista de bandas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banda'
    post:
      tags: [bandas]
      summary: Crear perfil de banda
      operationId: crearBanda
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBandaRequest'
      responses:
        '201':
          description: Banda creada.
  /bandas/ciudad/{id}:
    get:
      tags: [bandas]
      summary: Listar bandas de una ciudad
      operationId: bandasPorCiudad
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Bandas de la ciudad indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banda'
  /bandas/{id}:
    get:
      tags: [bandas]
      summary: Obtener perfil de banda
      operationId: getBanda
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Perfil completo de la banda con repertorio y eventos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Banda'
        '404':
          $ref: '#/components/responses/Error404'
    patch:
      tags: [bandas]
      summary: Editar perfil de banda (propietario o ADMIN)
      operationId: editarBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBandaRequest'
      responses:
        '200':
          description: Banda actualizada.
        '403':
          $ref: '#/components/responses/Error403'
  /bandas/{id}/verificar:
    patch:
      tags: [bandas]
      summary: Verificar / desverificar banda (solo ADMIN)
      operationId: verificarBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: boolean
      responses:
        '200':
          description: Estado de verificación actualizado.
  /bandas/{id}/eventos:
    get:
      tags: [bandas]
      summary: Obtener todos los eventos de una banda
      operationId: eventosDeUnaBanda
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de eventos ordenada cronológicamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evento'
    post:
      tags: [bandas]
      summary: Añadir evento a la agenda de la banda
      operationId: crearEvento
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventoRequest'
      responses:
        '201':
          description: Evento creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
  /bandas/{id}/eventos/{eventoId}:
    put:
      tags: [bandas]
      summary: Actualizar evento de la banda (propietario o ADMIN)
      operationId: actualizarEvento
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: eventoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventoRequest'
      responses:
        '200':
          description: Evento actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
    delete:
      tags: [bandas]
      summary: Eliminar evento de la banda (propietario o ADMIN)
      operationId: eliminarEvento
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: eventoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Evento eliminado.
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
  /bandas/{id}/agenda/{anio}:
    get:
      tags: [bandas]
      summary: Agenda completa de una banda para un año
      description: |
        Combina los eventos propios de la banda (conciertos, actuaciones) con las
        procesiones en las que participa ese año, ordenados cronológicamente.
      operationId: agendaAnual
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: anio
          in: path
          required: true
          schema:
            type: integer
            example: 2025
      responses:
        '200':
          description: Agenda unificada del año.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgendaItem'

  # ── PROCESIONES ───────────────────────────
...
  /seguimientos/banda/{id}:
    get:
      tags: [seguimientos]
      summary: Estado de seguimiento de una banda
      operationId: estadoSeguimientoBanda
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Estado de seguimiento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeguimientoEstado'
    post:
      tags: [seguimientos]
      summary: Seguir banda
      operationId: seguirBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: Seguimiento registrado.
    delete:
      tags: [seguimientos]
      summary: Dejar de seguir banda
      operationId: dejarBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Seguimiento eliminado.
...
  /publicaciones/banda/{id}:
    get:
      tags: [publicaciones]
      summary: Publicaciones de una banda
      operationId: publicacionesPorBanda
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de publicaciones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publicacion'
...
  /admin/bandas/{id}/verificar:
    put:
      tags: [admin]
      summary: Verificar banda
      operationId: adminVerificarBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Banda verificada.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en `POST /auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador numérico del recurso.
  responses:
...
    Error403:
      description: Sin permisos — el usuario no tiene el rol necesario.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RolUsuario:
      type: string
      enum: [admin, cofrade, hermandad, banda]
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/RolUsuario'
        verificado:
          type: boolean
        ciudadResidencia:
          $ref: '#/components/schemas/Ciudad'
...
    Hermandad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        nombrePopular:
          type: string
        titulares:
          type: array
          items:
            type: string
        añoFundacion:
          type: integer
          nullable: true
        templo:
          type: string
        diaSalida:
          type: string
          nullable: true
        descripcion:
          type: string
          nullable: true
        imagenEscudo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'

    # ── Bandas ────────────────────────────────
    CreateBandaRequest:
      type: object
      required: [nombre, estiloMusical]
      properties:
        nombre:
          type: string
          example: "Banda de Música de la Esperanza"
        estiloMusical:
          type: string
          example: "Banda de Cornetas y Tambores"
        direccion:
          type: string
        localidad:
          type: string
        codigoPostal:
          type: string
        numeroComponentes:
          type: integer
        ciudadId:
          type: integer
        historia:
          type: string
    UpdateBandaRequest:
      type: object
      properties:
        nombre:
          type: string
        estiloMusical:
          type: string
        historia:
          type: string
        numeroComponentes:
          type: integer
        repertorioIds:
          type: array
          items:
            type: integer
          description: IDs de marchas para actualizar el repertorio completo.
    Banda:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        estiloMusical:
          type: string
        historia:
          type: string
          nullable: true
        numeroComponentes:
          type: integer
          nullable: true
        imagenLogo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'
        repertorio:
          type: array
          items:
            $ref: '#/components/schemas/Marcha'
        eventos:
          type: array
          items:
            $ref: '#/components/schemas/Evento'

    # ── Procesiones ───────────────────────────
...
    CreateEventoRequest:
      type: object
      required: [titulo, fechaHora, lugar, bandaId]
      properties:
        titulo:
          type: string
          minLength: 5
          example: "Concierto de Primavera 2025"
        fechaHora:
          type: string
          format: date-time
          example: "2025-05-10T20:00:00"
        lugar:
          type: string
          example: "Auditorio Municipal de Sevilla"
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          enum: [procesion, concierto]
          default: concierto
        bandaId:
          type: integer
    UpdateEventoRequest:
      type: object
      properties:
        titulo:
          type: string
        fechaHora:
          type: string
          format: date-time
        lugar:
          type: string
        descripcion:
          type: string
        tipo:
          type: string
    Evento:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        fechaHora:
          type: string
          format: date-time
        lugar:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
        bandaId:
          type: integer
    AgendaItem:
      type: object
      properties:
        fecha:
          type: string
          format: date-time
        tipo:
          type: string
          example: "Procesión"
        nombre:
          type: string
        lugar:
          type: string
        detalle:
          type: string
          nullable: true

    # ── Publicaciones ─────────────────────────
...
    Publicacion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        contenido:
          type: string
        tipo:
          type: string
          enum: [general, itinerario]
        fechaCreacion:
          type: string
          format: date-time
          readOnly: true
        autor:
          $ref: '#/components/schemas/Usuario'
        hermandad:
          $ref: '#/components/schemas/Hermandad'
          nullable: true
        banda:
          $ref: '#/components/schemas/Banda'
          nullable: true

    # ── Seguimientos ──────────────────────────
    SeguimientoEstado:
      type: object
      properties:
        sigues:
          type: boolean
          description: Si el usuario autenticado sigue esta entidad.
        seguidores:
          type: integer
          description: Número total de seguidores.

    # ── Search ────────────────────────────────
...
    Ciudad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        provincia:
          type: string
        pais:
          type: string
...
    Marcha:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        compositor:
          type: string
          nullable: true

    # ── Admin ─────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "La procesión no existe"
        error:
          type: string
          example: "Not Found"
security:
  - BearerAuth: []